Moves analytics to Shoebox Module
