+++ kifi-backend/modules/common/app/com/keepit/common/analytics/EventLog.scala
@@ -74,18 +74,6 @@ case class Event(
   serverVersion: String
 )
 
-class EventRepo @Inject() (eventStore: EventStore, mongoEventStore: MongoEventStore) {
-  def persistToS3(event: Event): Event = {
-    eventStore += (event.externalId -> event)
-    event
-  }
-  def persistToMongo(event: Event): Event = mongoEventStore.save(event)
-  def persist(event: Event): Event = {
-    persistToS3(event)
-    persistToMongo(event)
-  }
-}
-
 object Events {
   def serverVersion(implicit fortyTwoServices: FortyTwoServices) = fortyTwoServices.currentService + ":" + fortyTwoServices.currentVersion
