+++ kifi-backend/modules/common/test/com/keepit/common/net/HttpClientTest.scala
@@ -13,7 +13,7 @@ class HttpClientTest extends Specification {
 
   "HttpClient" should {
     "instantiated" in {
-      running(new DeprecatedEmptyApplication().withFakePersistEvent.withShoeboxServiceModule) {
+      running(new DeprecatedEmptyApplication().withShoeboxServiceModule) {
         /*val client = instance[HttpClient]
         client !== null
         client.longTimeout() !== null*/
