+++ kifi-backend/modules/search/test/com/keepit/search/index/ArticleIndexerTest.scala
@@ -79,7 +79,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
   }
 
   "ArticleIndexer" should {
-    "index indexable URIs" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule())(new IndexerScope {
+    "index indexable URIs" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule())(new IndexerScope {
 
       uri1 = fakeShoeboxServiceClient.saveURIs(uri1.withState(INACTIVE)).head
       uri2 = fakeShoeboxServiceClient.saveURIs(uri2.withState(INACTIVE)).head
@@ -111,7 +111,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       indexer.sequenceNumber.value === currentSeqNum
     })
 
-    "search documents (hits in contents)" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "search documents (hits in contents)" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       indexer.search("alldocs").size === 3
@@ -129,7 +129,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.head.id === uriIdArray(2)
     })
 
-    "search documents (hits in titles)" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "search documents (hits in titles)" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       var res = indexer.search("title1")
@@ -145,7 +145,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.head.id === uriIdArray(2)
     })
 
-    "search documents (hits in contents and titles)" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "search documents (hits in contents and titles)" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       var res = indexer.search("title1 alldocs")
@@ -161,7 +161,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.head.id === uriIdArray(2)
     })
 
-    "search documents using stemming" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "search documents using stemming" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       indexer.search("alldoc").size === 3
@@ -172,7 +172,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       indexer.search("+body +soul").size === 3
     })
 
-    "limit the result by percentMatch" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "limit the result by percentMatch" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       var res = indexer.search("title1 alldocs", percentMatch = 0.0f)
@@ -191,7 +191,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.size === 0
     })
 
-    "limit the result by site" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "limit the result by site" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       var res = indexer.search("alldocs")
@@ -222,7 +222,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.size === 2
     })
 
-    "match on the URI" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "match on the URI" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       var res = indexer.search("keepit")
@@ -235,7 +235,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       res.size === 1
     })
 
-    "be able to dump Lucene Document" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "be able to dump Lucene Document" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
 
       store += (uri1.id.get -> mkArticle(uri1.id.get, "title1 titles", "content1 alldocs body soul"))
@@ -244,7 +244,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       doc.getFields.forall{ f => indexer.getFieldDecoder(f.name).apply(f).length > 0 } === true
     })
 
-    "delete documents with inactive, active, unscrapable, oe scrape_wanted state" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "delete documents with inactive, active, unscrapable, oe scrape_wanted state" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       indexer.run()
       indexer.numDocs === 3
 
@@ -285,7 +285,7 @@ class ArticleIndexerTest extends Specification with ApplicationInjector {
       indexer.search("content3").size === 1
     })
 
-    "retrieve article records from index" in running(new DeprecatedEmptyApplication().withFakePersistEvent().withShoeboxServiceModule)(new IndexerScope {
+    "retrieve article records from index" in running(new DeprecatedEmptyApplication().withShoeboxServiceModule)(new IndexerScope {
       import com.keepit.search.index.ArticleRecordSerializer._
       indexer.run()
       indexer.numDocs === 3
